•	Windows and Linux Authentication and Centralization login profile. (Samba Server with Windows 7 client) (July 2014 – August 2014)
Company: Hewlett Packard
Abstract: Samba is a suite of programs that provide seamless file and
print services for Windows clients in Linux. Unlike other
implementations of the SMB/CIFS networking protocol, Samba is freely
available, and allows for interoperability between Linux/Unix servers
and Windows/Unix/Linux clients. The project purpose was to create a
domain based environment for which a Linux System was used as the
server, on which Samba Server was installed and windows systems were the
clients. Thus, the profiles were created on the server which could be
used by windows system for login purpose and data would be stored on the
server.
Language: Linux Administration with Shell Scripting